printf bon

+++ libft/src/ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: rgermain <marvin@le-101.fr>                +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2018/08/02 11:14:09 by rgermain     #+#   ##    ##    #+#       */
-/*   Updated: 2018/11/11 15:26:42 by rgermain    ###    #+. /#+    ###.fr     */
+/*   Updated: 2018/11/01 18:27:26 by rgermain    ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
@@ -15,9 +15,9 @@
 
 int	ft_atoi(const char *str)
 {
-	long	a;
-	long	b;
-	long	neg;
+	int	a;
+	int	b;
+	int	neg;
 
 	a = 0;
 	b = 0;
@@ -35,5 +35,5 @@ int	ft_atoi(const char *str)
 		b = ((b * 10) + (str[a] - '0'));
 		a++;
 	}
-	return ((int)(b * neg));
+	return (b * neg);
 }
